requirements.txt
# Requirements for Food Calorie Prediction Project

# Deep Learning
torch>=1.10.0
torchvision>=0.11.0

# Data Processing
pandas>=1.3.0
numpy>=1.21.0
pillow>=8.0.0

# Augmentation
albumentations>=1.1.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Utilities
tqdm>=4.62.0
pyyaml>=5.4.0
scikit-learn>=1.0.0

# Optional: Jupyter
jupyter>=1.0.0
notebook>=6.4.0
ipywidgets>=7.6.0
